employee_management_system/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
pydantic==2.5.0


# Frontend requirements
streamlit==1.29.0
requests==2.31.0
pandas==2.1.3
python-dotenv==1.0.0
pytest==7.4.3